fix: rename faulty import path after refactor
